fix(opensmt/api/opensmt_c): opensmt_mk_int_var takes long instead of int
close #67

Note: dReal uses a floating-point interval as an internal
representation. That is, the long integer will be casted to double and
used inside of dReal.

+++ src/opensmt/api/opensmt_c.C
@@ -278,7 +278,7 @@ opensmt_expr opensmt_mk_bool_var( opensmt_context c, char * s )
   return static_cast< void * >( res );
 }
 
-opensmt_expr opensmt_mk_int_var( opensmt_context c, char * s , int lb, int ub)
+opensmt_expr opensmt_mk_int_var( opensmt_context c, char * s , long lb, long ub)
 {
   assert( c );
   assert( s );
