+++ src/opensmt/api/opensmt_c.h
@@ -73,7 +73,7 @@ void             opensmt_print_interpolant         ( opensmt_context, const char
 opensmt_expr     opensmt_mk_true                   ( opensmt_context );
 opensmt_expr     opensmt_mk_false                  ( opensmt_context );
 opensmt_expr     opensmt_mk_bool_var               ( opensmt_context, char * );
-opensmt_expr     opensmt_mk_int_var                ( opensmt_context, char * , int , int );
+opensmt_expr     opensmt_mk_int_var                ( opensmt_context, char * , long , long );
 opensmt_expr     opensmt_mk_real_var               ( opensmt_context, char * , double, double );
 opensmt_expr     opensmt_mk_or                     ( opensmt_context, opensmt_expr *, unsigned );
 opensmt_expr     opensmt_mk_and                    ( opensmt_context, opensmt_expr *, unsigned );
